fix: [007-GROUP/UPDATE] 사진 삭제/변경없음 처리 구분, 기본 사진 삭제되는 것 방지

+++ src/main/java/com/valuewith/tweaver/defaultImage/service/ImageService.java
@@ -117,7 +117,11 @@ public class ImageService {
                 throw new S3ImageNotFoundException(ErrorCode.S3_IMAGE_NOT_FOUND);
             }
 
-            amazonS3.deleteObject(bucketName, currentKey);
+            boolean exists = defaultImageRepository.existsDefaultImageByImageName(currentUrl);
+            if(!exists) {
+                amazonS3.deleteObject(bucketName, currentKey);
+            }
+
             return newImageUrl;
         } catch (Exception e) {
             String newKey = generateKey(newImageUrl);
